docker_project/requirements.txt
selenium
beautifulsoup4
webdriver-manager
lxml
pymongo[srv]
datetime
python-dotenv
groq
